+++ Changelog.md
@@ -5,3 +5,11 @@ The problem of uppercases characters in the installation script has been solved 
 ---------------------------------------------19 / 04 / 2016----------------------------------------------------
 
 Removed Inguma from the menu to bypass issue with installation (By Jayson Grace : https://github.com/l50)
+
+---------------------------------------------17 / 05 / 2022----------------------------------------------------
+
+Updated the script to work with Python 3.10 and above, Python 2 is no longer supported
+
+---------------------------------------------29 / 09 / 2022----------------------------------------------------
+
+New splash for the script, reformat the code and updated some functions and variables names.
