+++ todo.md
@@ -1,0 +1,18 @@
+Install.sh:
+    - Check Package Manager
+    - Check Dependencies
+    - Check Configuration
+    - Check Permissions
+    - Display Welcome Message
+    - Install Packages
+    - Install Dependencies
+    - Install Configuration
+    - Display Warning Message about the repository
+    - Adding GPG Key
+
+katoolin.py:
+    - Update Katoolin to python3
+    - Fix GPG Key Error
+    - Fix Katoolin Error
+    - Detest Linux Distribution
+    - Install Katoolin on any Linux Distribution
